backend/requirements.txt
numpy
matplotlib
tensorflow
opencv-python
scikit-learn
flask-cors

